Fully move paraview extractor to conda_package
This allows users to call it as a function, rather than just as
a command-line tool.
