+++ conda_package/mpas_tools/__init__.py
@@ -1,2 +1,2 @@
-__version_info__ = (0, 0, 9)
+__version_info__ = (0, 0, 10)
 __version__ = '.'.join(str(vi) for vi in __version_info__)
