+++ conda_package/mpas_tools/viz/paraview_extractor.py
@@ -1,10 +1,71 @@
-#!/usr/bin/env python
 """
-Name: viz.py
-Authors: Xylar Asay-Davis
+Extract a field from a time series of NetCDF files as VTK files for plotting in
+paraview.
 
-Utility library for various scripts used to extract vtk geometry from NetCDF
-files on MPAS grids.
+It can extract a field across multiple files by passing in a regular expression
+for the filename patter. As an example, one can run the script using:
+
+    ./paraview_vtk_field_extractor.py -v areaCell,latVertex -f "hist.comp.*.nc"
+
+To extract a time series of areaCell,latVertex that spans multiple files.
+By default, time-independent fields on cells are written to a file
+vtk_files/staticFieldsOnCells.vtp
+and time-dependent fields on cells are written to
+vtk_files/timeDependentFieldsOnCells.pvd
+vtk_files/time_series/timeDependentFieldsOnCells.0.vtp
+vtk_files/time_series/timeDependentFieldsOnCells.1.vtp
+...
+and similarly for edges and vertices.  Time-independent fields can be
+included in each time step of the time-dependent fields for with
+the --combine flag.  This allows time-dependent and -independent fields
+to be combined in filters within Paraview at the expense of considerable
+additional storage space.
+
+Indices for extra dimensions can either be supplied at runtime at a prompt
+or through the -d flag.  For each extra dimension, the use can specifiy
+nothing (an empty string, meaning skip any fields with this dimension),
+a single index, or a comma-separated list of indices or a range of indices
+indices (separated by 1 or 2 colons).  For example,
+
+    -d maxEdges= nVertLeves=0:10:2 nParticles=0,2,4,6,8
+
+will ignore any fields with dimension maxEdges, extract every other layer from
+the first 10 vertical levels (each into its own field) and extract the five
+specified particles.
+
+An index array can also be specified in this way (and these can be mixed with
+integer indices in a comma-separated list but not in a colon-separated range):
+
+    -d nVertLeves=0,maxLevelCell
+
+will extract fields from the first vertical level and the vertical level with
+index given by maxLevelCell.
+
+The extractor includes optional support for extracting geometry appropriate
+for displaying variables at the depth of a topographic feature (typically the
+top or bottom of the domain) for MPAS components with a spatially variable
+top or bottom index (e.g. `maxLevelCell` in MPAS-Ocean).  This is accomplished
+with flags such as:
+
+    --topo_dim=nVertLevels --topo_cell_index=maxLevelCell
+
+Fields on cells are sampled at the topographic index and the geometry includes
+polygons corresponding to edges so that vertical faces between adjacent cells
+can be displayed.  Fields are extracted as normal except that they are sampled
+as point data rather than cell data, allowing computations in ParaView to
+display the topography.  A mask field is also included indicating which parts
+of edge polygons correspond to the boundary of the domain (boundaryMask == 1)
+and which parts of cell and edge polygons are interior (boundaryMask == 0).
+Together, this can be used to plot topography by using a calculator filter like
+the following:
+
+    coords*(1.0 + 100.0/mag(coords)*((1 - boundaryMask)*(-bottomDepth)
+                                     + 10.0*boundaryMask))
+
+If this is entered into a Calculator Filter in ParaView with the "coordinate
+result" box checked, the result will to display the MPAS-Ocean topography,
+exaggerated by a factor of 100, with a value equivalent to 10 m along boundary
+points of edge polygons (a "water-tight" surface).
 """
 
 from __future__ import absolute_import, division, print_function, \
@@ -13,8 +74,10 @@ from __future__ import absolute_import, division, print_function, \
 from pyevtk.vtk import VtkFile, VtkPolyData
 
 import sys
+import os
 import glob
 import numpy
+import argparse
 
 from builtins import input
 
@@ -25,6 +88,629 @@ try:
     use_progress_bar = True
 except ImportError:
     use_progress_bar = False
+
+
+def extract_vtk(filename_pattern, variable_list='all', dimension_list=None,
+                mesh_filename=None, blocking=10000, output_32bit=False,
+                combine=False, append=False, out_dir=None, xtime='xtime',
+                lonlat=False, time=None, ignore_time=False, topo_dim=None,
+                topo_cell_index=None):
+    """
+    Extract fields from a time series of NetCDF files as VTK files for plotting
+    in paraview.
+
+    To extract fields across multiple files, passing in a regular expression
+    for the filename pattern, for exmaple ``filename_pattern="hist.comp.*.nc"``.
+
+    By default, time-independent fields on cells are written to a file
+    
+    .. code-block::
+
+      vtk_files/staticFieldsOnCells.vtp
+
+    and time-dependent fields on cells are written to
+    
+    .. code-block::
+
+      vtk_files/timeDependentFieldsOnCells.pvd
+      vtk_files/time_series/timeDependentFieldsOnCells.0.vtp
+      vtk_files/time_series/timeDependentFieldsOnCells.1.vtp
+      ...
+
+    and similarly for edges and vertices.  Time-independent fields can be
+    included in each time step of the time-dependent fields for with
+    ``combine=True``.  This allows time-dependent and -independent fields
+    to be combined in filters within Paraview at the expense of considerable
+    additional storage space.
+
+    Indices for extra dimensions can either be supplied at runtime at a prompt
+    (if ``dimension_list=None``) or via a list of strings with the dimensions
+    and associated indices.  For each extra dimension, you can specifiy
+    nothing for the indices (an empty string, meaning skip any fields with this
+    dimension), a single index, a comma-separated list of indices, or a range
+    of indices (separated by 1 or 2 colons).  For example,
+
+    .. code-block::
+
+      dimension_list=['maxEdges=', 'nVertLeves=0:10:2', 'nParticles=0,2,4,6,8']
+
+    will ignore any fields with dimension ``maxEdges``, extract every other
+    layer from the first 10 vertical levels (each into its own field) and
+    extract the 5 specified particles.
+
+    An index array can also be specified in this way (and these can be mixed
+    with integer indices in a comma-separated list but not in a colon-separated
+    range):
+
+    .. code-block::
+
+      dimension_list=['nVertLeves=0,maxLevelCell']
+
+    will extract fields from the first vertical level and the vertical level
+    with index given by ``maxLevelCell``.
+
+    The extractor includes optional support for extracting geometry appropriate
+    for displaying variables at the depth of a topographic feature (typically
+    the top or bottom of the domain) for MPAS components with a spatially
+    variable top or bottom index (e.g. ``maxLevelCell`` in MPAS-Ocean).  This is
+    accomplished with arguments such as:
+
+    .. code-block::
+
+      topo_dim='nVertLevels', topo_cell_index='maxLevelCell'
+
+    Fields on cells are sampled at the topographic index and the geometry
+    includes polygons corresponding to edges so that vertical faces between
+    adjacent cells can be displayed.  Fields are extracted as normal except
+    that they are sampled as point data rather than cell data, allowing
+    computations in ParaView to display the topography.  A mask field is also
+    included indicating which parts of edge polygons correspond to the boundary
+    of the domain (``boundaryMask == 1``) and which parts of cell and edge
+    polygons are interior (``boundaryMask == 0``).  Together, this can be used
+    to plot topography by using a calculator filter like the following:
+
+    .. code-block::
+
+      coords*(1.0 + 100.0/mag(coords)*((1 - boundaryMask)*(-bottomDepth)
+                                       + 10.0*boundaryMask))
+
+    If this is entered into a Calculator Filter in ParaView with the "coordinate
+    result" box checked, the result will to display the MPAS-Ocean topography,
+    exaggerated by a factor of 100, with a value equivalent to 10 m along
+    boundary points of edge polygons (a "water-tight" surface).
+
+    Parameters
+    ----------
+    filename_pattern : str
+        MPAS Filename pattern
+
+    variable_list: list of str, optional
+        List of variables to extract ('all' for all variables, 'allOnCells'
+        for all variables on cells, etc.)"
+
+    dimension_list: list of str, optional
+        A list of dimensions and associated indices
+
+    mesh_filename: str, optional
+        MPAS Mesh filename. By default, the first file matching
+        ``filename_pattern`` will be used
+
+    blocking: int, optional
+        Size of blocks when reading MPAS file
+
+    output_32bit: bool, optional
+        Whether the vtk files will be written using 32bit floats
+
+    combine: bool, optional
+        Whether time-independent fields are written to each file along with
+        time-dependent fields
+
+    append: bool, optional
+        Whether only vtp files that do not already exist are written out
+
+    out_dir: str, optional
+        The output directory
+
+    xtime: str, optional"
+        The name of the xtime variable or 'none' to extract Time dim without
+        xtime
+
+    lonlat: bool, optional
+        Whether the resulting points are in lon-lat space, not Cartesian
+
+    time: str, optional
+        Indices for the time dimension
+
+    ignore_time: bool, optional
+        Whether to ignore the Time dimension if it exists for files with a Time
+        dimension but no xtime variable (e.g. mesh file)
+
+    topo_dim: str, optional
+        Dimension and range for topography dimension
+
+    topo_cell_index: str, optional
+        Index array indicating the bottom of the domain (default is the
+        topo_dim-1 for all cells)
+    """
+
+    if ignore_time:
+        xtime = None
+
+    (time_indices, time_file_names) = setup_time_indices(
+        filename_pattern, xtime)
+
+    if time is not None:
+        time_indices, time_file_names = \
+            parse_time_indices(time, time_indices, time_file_names)
+
+    separate_mesh_file = True
+    if mesh_filename is None:
+        mesh_filename = time_file_names[0]
+        separate_mesh_file = False
+
+    # Setting dimension values:
+    time_series_file = open_netcdf(time_file_names[0])
+    if separate_mesh_file:
+        mesh_file = open_netcdf(mesh_filename)
+    else:
+        mesh_file = None
+    extra_dims, topo_cell_indices = \
+        parse_extra_dims(dimension_list, time_series_file,
+                         mesh_file, topo_dim=topo_dim,
+                         topo_cell_index_name=topo_cell_index)
+    basic_dims = ['nCells', 'nEdges', 'nVertices', 'Time']
+    include_dims = ['nCells', 'nEdges', 'nVertices']
+    if topo_dim is not None:
+        basic_dims.append(topo_dim)
+        include_dims = ['nCells']
+
+    (all_dim_vals, cellVars, vertexVars, edgeVars) = \
+        setup_dimension_values_and_sort_vars(
+            time_series_file, mesh_file,  variable_list, extra_dims,
+            basic_dims=basic_dims)
+    time_series_file.close()
+    if(mesh_file is not None):
+        mesh_file.close()
+
+    summarize_extraction(mesh_filename, time_indices, cellVars,
+                         vertexVars, edgeVars)
+
+    # Handle cell variables
+    if len(cellVars) > 0:
+        print(" -- Extracting cell fields --")
+
+        mesh_file = open_netcdf(mesh_filename)
+
+        # Build cell geometry
+        if topo_dim is None:
+            (vertices, connectivity, offsets, valid_mask) = \
+                build_cell_geom_lists(mesh_file, output_32bit, lonlat)
+            cell_to_point_map = None
+            boundary_mask = None
+        else:
+            (vertices, connectivity, offsets, valid_mask, cell_to_point_map,
+             boundary_mask) = build_topo_point_and_polygon_lists(
+                     mesh_file, output_32bit, lonlat)
+
+        if not separate_mesh_file:
+            mesh_file.close()
+            mesh_file = None
+
+        build_field_time_series(time_indices, time_file_names, mesh_file,
+                                out_dir, blocking, all_dim_vals,
+                                'nCells', cellVars, vertices, connectivity,
+                                offsets, valid_mask, output_32bit,
+                                combine, append, xtime,
+                                topo_dim=topo_dim,
+                                topo_cell_indices=topo_cell_indices,
+                                cell_to_point_map=cell_to_point_map,
+                                boundary_mask=boundary_mask)
+        if separate_mesh_file:
+            mesh_file.close()
+
+        print("")
+
+    if len(vertexVars) > 0:
+        print(" -- Extracting vertex fields --")
+
+        mesh_file = open_netcdf(mesh_filename)
+
+        # Build vertex geometry
+        (vertices, connectivity, offsets, valid_mask) = \
+            build_vertex_geom_lists(mesh_file, output_32bit, lonlat)
+
+        if not separate_mesh_file:
+            mesh_file.close()
+            mesh_file = None
+
+        build_field_time_series(time_indices, time_file_names, mesh_file,
+                                out_dir, blocking, all_dim_vals,
+                                'nVertices', vertexVars, vertices,
+                                connectivity, offsets, valid_mask, output_32bit,
+                                combine, append, xtime)
+
+        if separate_mesh_file:
+            mesh_file.close()
+
+        print("")
+
+    if len(edgeVars) > 0:
+        print(" -- Extracting edge fields --")
+
+        mesh_file = open_netcdf(mesh_filename)
+
+        # Build cell list
+        (vertices, connectivity, offsets, valid_mask) = \
+            build_edge_geom_lists(mesh_file, output_32bit, lonlat)
+
+        if not separate_mesh_file:
+            mesh_file.close()
+            mesh_file = None
+
+        build_field_time_series(time_indices, time_file_names, mesh_file,
+                                out_dir, blocking, all_dim_vals,
+                                'nEdges', edgeVars, vertices, connectivity,
+                                offsets, valid_mask, output_32bit,
+                                combine, append, xtime)
+
+        if separate_mesh_file:
+            mesh_file.close()
+
+
+def main():
+    if use_progress_bar:
+        print(" -- Using progress bars --")
+    else:
+        print(" -- Progress bars are not available--")
+    parser = \
+        argparse.ArgumentParser(description=__doc__,
+                                formatter_class=argparse.RawTextHelpFormatter)
+    parser.add_argument("-f", "--file_pattern", dest="filename_pattern",
+                        help="MPAS Filename pattern.", metavar="FILE",
+                        required=True)
+    parser.add_argument("-m", "--mesh_file", dest="mesh_filename",
+                        help="MPAS Mesh filename. If not set, it will use the "
+                             "first file in the -f flag as the mesh file.")
+    parser.add_argument("-b", "--blocking", dest="blocking", type=int,
+                        help="Size of blocks when reading MPAS file",
+                        metavar="BLK", default=32000)
+    parser.add_argument("-v", "--variable_list", dest="variable_list",
+                        help="List of variables to extract ('all' for all "
+                             "variables, 'allOnCells' for all variables on "
+                             "cells, etc.)", metavar="VAR", required=True)
+    parser.add_argument("-3", "--32bit", dest="output_32bit",
+                        help="If set, the vtk files will be written using "
+                             "32bit floats.", action="store_true")
+    parser.add_argument("-c", "--combine", dest="combine_output",
+                        help="If set, time-independent fields are written to "
+                             "each file along with time-dependent fields.",
+                        action="store_true")
+    parser.add_argument("-a", "--append", dest="append",
+                        help="If set, only vtp files that do not already "
+                             "exist are written out.", action="store_true")
+    parser.add_argument("-d", "--dim_list", dest="dimension_list", nargs="+",
+                        help="A list of dimensions and associated indices.",
+                        metavar="DIM")
+    parser.add_argument("-o", "--out_dir", dest="out_dir",
+                        help="the output directory.", default='vtk_files',
+                        metavar="DIR")
+    parser.add_argument("-x", "--xtime", dest="xtime", default='xtime',
+                        metavar="XTIME",
+                        help="the name of the xtime variable or 'none' to "
+                             "extract Time dim without xtime")
+    parser.add_argument("-l", "--lonlat", dest="lonlat",
+                        help="If set, the resulting points are in lon-lat "
+                             "space, not Cartesian.", action="store_true")
+    parser.add_argument("-t", "--time", dest="time",
+                        help="Indices for the time dimension", metavar="TIME",
+                        required=False)
+    parser.add_argument("--ignore_time", dest="ignore_time", required=False,
+                        action="store_true",
+                        help="ignore the Time dimension if it exists "
+                             "for files with a Time dimension but no xtime"
+                             "variable (e.g. mesh file)")
+    parser.add_argument("--topo_dim", dest="topo_dim", required=False,
+                        help="Dimension and range for topography dimension")
+    parser.add_argument("--topo_cell_index", dest="topo_cell_index",
+                        required=False,
+                        help="Index array indicating the bottom of the domain "
+                             "(default is the topo_dim-1 for all cells)")
+    args = parser.parse_args()
+
+    extract_vtk(filename_pattern=args.filename_pattern, 
+                variable_list=args.variable_list,
+                dimension_list=args.dimension_list,
+                mesh_filename=args.mesh_filename,
+                blocking=args.blocking,
+                output_32bit=args.output_32bit,
+                combine=args.combine_output,
+                append=args.append,
+                out_dir=args.out_dir,
+                xtime=args.xtime,
+                lonlat=args.lonlat,
+                time=args.time,
+                ignore_time=args.ignore_time,
+                topo_dim=args.topo_dim,
+                topo_cell_index=args.topo_cell_index)
+
+
+def build_field_time_series(local_time_indices, file_names, mesh_file,
+                            out_dir, blocking, all_dim_vals, blockDimName,
+                            variable_list, vertices, connectivity, offsets,
+                            valid_mask, output_32bit, combine_output, append,
+                            xtimeName, topo_dim=None, topo_cell_indices=None,
+                            cell_to_point_map=None, boundary_mask=None):  # {{{
+
+    if len(variable_list) == 0:
+        return
+
+    if output_32bit:
+        outType = 'float32'
+    else:
+        outType = 'float64'
+
+    # Get dimension info to allocate the size of Colors
+    time_series_file = open_netcdf(file_names[0])
+
+    if mesh_file is not None:
+        # blockDim may not exist in time series file
+        blockDim = len(mesh_file.dimensions[blockDimName])
+    else:
+        blockDim = len(time_series_file.dimensions[blockDimName])
+
+    if boundary_mask is not None:
+        variable_list.append('boundaryMask')
+        all_dim_vals['boundaryMask'] = None
+        pointData = True
+        cellData = False
+    else:
+        pointData = False
+        cellData = True
+
+    # Pre-compute the number of blocks
+    nBlocks = int(numpy.ceil(blockDim / blocking))
+
+    nPolygons = len(offsets)
+    nPoints = len(vertices[0])
+    nTimes = len(local_time_indices)
+    nVars = len(variable_list)
+
+    var_has_time_dim = numpy.zeros(nVars, bool)
+    nHyperSlabs = 0
+    for iVar in range(nVars):
+        var_name = variable_list[iVar]
+        if boundary_mask is not None and var_name == 'boundaryMask':
+            var_has_time_dim[iVar] = False
+        elif xtimeName is not None:
+            if var_name in time_series_file.variables:
+                var_has_time_dim[iVar] = \
+                    'Time' in time_series_file.variables[var_name].dimensions
+            else:
+                # we can't support time dependence in the mesh file
+                assert('Time' not in mesh_file.variables[var_name].dimensions)
+                var_has_time_dim[iVar] = False
+
+        extra_dim_vals = all_dim_vals[var_name]
+        if (extra_dim_vals is None) or (len(extra_dim_vals) == 0):
+            nHyperSlabs += 1
+        else:
+            nHyperSlabs += len(extra_dim_vals)
+
+    any_var_has_time_dim = numpy.any(var_has_time_dim)
+
+    if topo_dim is not None:
+        if (mesh_file is not None) and (topo_dim in mesh_file.dimensions):
+            nTopoLevels = len(mesh_file.dimensions[topo_dim])
+        else:
+            nTopoLevels = len(time_series_file.dimensions[topo_dim])
+    else:
+        nTopoLevels = None
+
+    time_series_file.close()
+
+    try:
+        os.makedirs(out_dir)
+    except OSError:
+        pass
+
+    if any_var_has_time_dim:
+        try:
+            os.makedirs('{}/time_series'.format(out_dir))
+        except OSError:
+            pass
+    else:
+        # there is no point in combining output if no fields have Time dim
+        combine_output = False
+        nTimes = 1
+
+    # Output time series
+    if use_progress_bar:
+        widgets = ['Writing time series: ', Percentage(), ' ', Bar(), ' ',
+                   ETA()]
+        field_bar = ProgressBar(widgets=widgets,
+                                maxval=nTimes*nHyperSlabs).start()
+    else:
+        print("Writing time series....")
+
+    suffix = blockDimName[1:]
+    if any_var_has_time_dim:
+        if combine_output or numpy.all(var_has_time_dim):
+            out_prefix = "fieldsOn{}".format(suffix)
+        else:
+            out_prefix = "timeDependentFieldsOn{}".format(suffix)
+        # start the pvd file
+        pvd_file = write_pvd_header(out_dir, out_prefix)
+        pvd_file.write('<Collection>\n')
+
+    if not combine_output and not numpy.all(var_has_time_dim):
+        static_prefix = "staticFieldsOn{}".format(suffix)
+        varIndices = numpy.arange(nVars)[numpy.logical_not(var_has_time_dim)]
+        timeIndependentFile = write_vtp_header(out_dir,
+                                               static_prefix,
+                                               varIndices[0],
+                                               varIndices,
+                                               variable_list,
+                                               all_dim_vals,
+                                               vertices,
+                                               connectivity,
+                                               offsets,
+                                               nPoints,
+                                               nPolygons,
+                                               outType,
+                                               cellData=cellData,
+                                               pointData=pointData,
+                                               xtime=None)
+
+    prev_file = ""
+    for time_index in range(nTimes):
+
+        if prev_file != file_names[time_index]:
+            if prev_file != "":
+                time_series_file.close()
+            time_series_file = open_netcdf(file_names[time_index])
+            prev_file = file_names[time_index]
+
+        if any_var_has_time_dim:
+            if xtimeName is None:
+                xtime = None
+                years = float(time_index)
+            else:
+                if xtimeName == 'none':
+                    xtime = '{}'.format(time_index)
+                    years = float(time_index)
+                else:
+                    if xtimeName not in time_series_file.variables:
+                        raise ValueError("xtime variable name {} not found in "
+                                         "{}".format(xtimeName,
+                                                     time_series_file))
+                    var = time_series_file.variables[xtimeName]
+                    if len(var.shape) == 2:
+                        xtime = var[local_time_indices[time_index],
+                                    :].tostring().decode('utf-8').strip()
+                        date = datetime(int(xtime[0:4]), int(xtime[5:7]),
+                                        int(xtime[8:10]), int(xtime[11:13]),
+                                        int(xtime[14:16]), int(xtime[17:19]))
+                        years = date2num(date, units='days since 0000-01-01',
+                                         calendar='noleap')/365.
+                    else:
+                        xtime = var[local_time_indices[time_index]]
+                        years = xtime/365.
+                        xtime = str(xtime)
+
+            # write the header for the vtp file
+            vtp_file_prefix = "time_series/{}.{:d}".format(out_prefix,
+                                                           time_index)
+            file_name = '{}/{}.vtp'.format(out_dir, vtp_file_prefix)
+            if append and os.path.exists(file_name):
+                pvd_file.write('<DataSet timestep="{:.16f}" group="" '
+                               'part="0"\n'.format(years))
+                pvd_file.write('\tfile="{}.vtp"/>\n'.format(vtp_file_prefix))
+                continue
+
+            if combine_output:
+                varIndices = numpy.arange(nVars)
+            else:
+                varIndices = numpy.arange(nVars)[var_has_time_dim]
+            timeDependentFile = write_vtp_header(out_dir,
+                                                 vtp_file_prefix,
+                                                 varIndices[0],
+                                                 varIndices,
+                                                 variable_list,
+                                                 all_dim_vals,
+                                                 vertices,
+                                                 connectivity,
+                                                 offsets,
+                                                 nPoints,
+                                                 nPolygons,
+                                                 outType,
+                                                 cellData=cellData,
+                                                 pointData=pointData,
+                                                 xtime=xtime)
+
+            # add time step to pdv file
+            pvd_file.write('<DataSet timestep="{:.16f}" group="" '
+                           'part="0"\n'.format(years))
+            pvd_file.write('\tfile="{}.vtp"/>\n'.format(vtp_file_prefix))
+
+        if time_index == 0 or combine_output:
+            varIndices = numpy.arange(nVars)
+        else:
+            # only the time-dependent variables
+            varIndices = numpy.arange(nVars)[var_has_time_dim]
+
+        iHyperSlabProgress = 0
+        for iVar in varIndices:
+            has_time = var_has_time_dim[iVar]
+
+            var_name = variable_list[iVar]
+            (out_var_names, dim_list) = \
+                get_hyperslab_name_and_dims(var_name, all_dim_vals[var_name])
+            if has_time or combine_output:
+                vtkFile = timeDependentFile
+            else:
+                vtkFile = timeIndependentFile
+            for iHyperSlab in range(len(out_var_names)):
+                if dim_list is not None:
+                    dim_vals = dim_list[iHyperSlab]
+                else:
+                    dim_vals = None
+
+                if boundary_mask is not None and var_name == 'boundaryMask':
+                    field = numpy.array(boundary_mask, dtype=outType)
+                else:
+                    field = numpy.zeros(blockDim, dtype=outType)
+
+                    for iBlock in numpy.arange(0, nBlocks):
+                        blockStart = iBlock * blocking
+                        blockEnd = min((iBlock + 1) * blocking, blockDim)
+                        block_indices = numpy.arange(blockStart, blockEnd)
+                        if topo_cell_indices is None:
+                            block_topo_cell_indices = None
+                        else:
+                            block_topo_cell_indices = \
+                                topo_cell_indices[block_indices]
+                        field_block = read_field(
+                            var_name, mesh_file, time_series_file,
+                            dim_vals, local_time_indices[time_index],
+                            block_indices, outType,  topo_dim=topo_dim,
+                            topo_cell_indices=block_topo_cell_indices,
+                            nTopoLevels=nTopoLevels)
+
+                        field[blockStart:blockEnd] = field_block
+
+                    field = field[valid_mask]
+
+                    if cell_to_point_map is not None:
+                        # map field from cells to points
+                        field = field[cell_to_point_map]
+
+                vtkFile.appendData(field)
+
+                if use_progress_bar:
+                    field_bar.update(time_index*nHyperSlabs +
+                                     iHyperSlabProgress)
+                    iHyperSlabProgress += 1
+
+                del field
+
+        if any_var_has_time_dim:
+            timeDependentFile.save()
+            del timeDependentFile
+
+        if time_index == 0 and not combine_output and not \
+                numpy.all(var_has_time_dim):
+            timeIndependentFile.save()
+            del timeIndependentFile
+
+    time_series_file.close()
+    if use_progress_bar:
+        field_bar.finish()
+
+    if any_var_has_time_dim:
+        # finish the pdv file
+        pvd_file.write('</Collection>\n')
+        pvd_file.write('</VTKFile>\n')
+        pvd_file.close()  # }}}
 
 
 def open_netcdf(file_name):
@@ -217,7 +903,7 @@ def parse_time_indices(index_string, time_indices, time_file_names):  # {{{
 
     Parameters
     ----------
-    index_string : str
+    index_string : str or list of int
         An index string indicating with indices are to be extracted
 
     time_indices : list of int
@@ -236,7 +922,7 @@ def parse_time_indices(index_string, time_indices, time_file_names):  # {{{
 
     """
     dim_size = len(time_indices)
-    indices, numerical_indices = parse_index_string(index_string, dim_size)
+    _, numerical_indices = parse_index_string(index_string, dim_size)
 
     time_indices = [time_indices[index] for index in numerical_indices]
     time_file_names = [time_file_names[index] for index in numerical_indices]
@@ -258,7 +944,7 @@ def parse_index_string(index_string, dim_size):  # {{{
 
     Parameters
     ----------
-    index_string : str
+    index_string : str or list of int
         An index string indicating with indices are to be extracted
 
     dim_size : int
@@ -269,44 +955,55 @@ def parse_index_string(index_string, dim_size):  # {{{
     indices : list of int
         The indices corresponding to the given index string.
     """
-    if index_string == '':
-        return [], []
-
-    if ',' in index_string:
-        indices = [index for index in index_string.split(',')]
-    elif ':' in index_string:
-        index_list = index_string.split(':')
-        if len(index_list) not in [2, 3]:
-            raise ValueError("Improperly formatted index string: "
-                             "{}".format(index_string))
-        if index_list[0] == '':
-            first = 0
+    if not isinstance(index_string, str):
+        numerical_indices = index_string 
+        indices = []
+        for index in numerical_indices:
+            if val < 0 or val >= dim_size:
+                raise ValueError("Index (or indices) out of bounds 0 <= "
+                                 "index < {}: {}".format(dim_size,
+                                                         index_string))
+            indices.append('{}'.format(index))
+    else:       
+        if index_string == '':
+            return [], []
+    
+        if ',' in index_string:
+            indices = [index for index in index_string.split(',')]
+        elif ':' in index_string:
+            index_list = index_string.split(':')
+            if len(index_list) not in [2, 3]:
+                raise ValueError("Improperly formatted index string: "
+                                 "{}".format(index_string))
+            if index_list[0] == '':
+                first = 0
+            else:
+                first = int(index_list[0])
+            if index_list[1] == '':
+                last = dim_size
+            else:
+                last = int(index_list[1])
+            if (len(index_list) == 2) or (index_list[2] == ''):
+                step = 1
+            else:
+                step = int(index_list[2])
+            indices = [str(index) for index in numpy.arange(first, last, step)]
         else:
-            first = int(index_list[0])
-        if index_list[1] == '':
-            last = dim_size
-        else:
-            last = int(index_list[1])
-        if (len(index_list) == 2) or (index_list[2] == ''):
-            step = 1
-        else:
-            step = int(index_list[2])
-        indices = [str(index) for index in numpy.arange(first, last, step)]
-    else:
-        indices = [index_string]
-
-    numerical_indices = []
-    for index in indices:
-        try:
-            val = int(index)
-        except ValueError:
-            continue
-
-        if val < 0 or val >= dim_size:
-            raise ValueError("Index (or indices) out of bounds 0 <= "
-                             "index < {}: {}".format(dim_size, index_string))
-
-        numerical_indices.append(val)
+            indices = [index_string]
+    
+        numerical_indices = []
+        for index in indices:
+            try:
+                val = int(index)
+            except ValueError:
+                continue
+    
+            if val < 0 or val >= dim_size:
+                raise ValueError("Index (or indices) out of bounds 0 <= "
+                                 "index < {}: {}".format(dim_size,
+                                                         index_string))
+    
+            numerical_indices.append(val)
 
     return indices, numerical_indices  # }}}
 
@@ -404,8 +1101,10 @@ def setup_dimension_values_and_sort_vars(
             for variable_name in mesh_file.variables:
                 add_var(mesh_file.variables, str(variable_name), include_dims,
                         exclude_dims)
-    else:
+    elif isinstance(variable_list, str):
         variable_names = variable_list.split(',')
+    else:
+        variable_names = variable_list
 
     for suffix in ['Cells', 'Edges', 'Vertices']:
         include_dim = 'n%s' % suffix
